Se implementa un método inicializador.
El método inicializador nos permite insertar datos en la BD, utiliza las
interfaces declaradas en el paquete de interfaces. Simulamos la
inserción de un cliente, un producto, una compra y agregamos detalles
sobre la misma.

+++ primera-entrega/src/main/java/com/coderhouse/entidades/Cliente.java
@@ -25,7 +25,7 @@ public class Cliente {
 	@Column(name = "cliente_id")
 	private Integer clienteID;
 	@Column(name = "telefono")
-	private Integer telefono;
+	private long telefono;
 	@Column(name = "nombre")
 	private String nombre;
 	@Column(name = "apellido")
@@ -54,11 +54,11 @@ public class Cliente {
 		this.clienteID = clienteID;
 	}
 
-	public Integer getTelefono() {
+	public Long getTelefono() {
 		return telefono;
 	}
 
-	public void setTelefono(Integer telefono) {
+	public void setTelefono(long telefono) {
 		this.telefono = telefono;
 	}
